Numerous changes as detailed below.
changes for latest commit

BASPHA.FOR increased damage for base phasers from 200 to 500
and removed the division by numply

DECWAR.FOR added code to implode game if etim == neg
this is needed because of etim bug

NOVA.FOR added ihita to NOVA.FOR for romulan handling. This had been
left out..so when a hit occured even though the amount was subtracted
from ship the figure shown as 0 hit. This fixes that problem.

PLNATK.FOR removed numply division from phaser hit computation
and strengthen base hits from 30 to 50 per build.

ROMDRV.FOR strengthened rom from random 200 + 200 to random 200 + 800
and increased romulan phaser hit from 200 to 300.
(big improvement!)

SETUP.FOR added message when high segment is initialized.

WARMAC.MAC added options in beginning of file for listing (LST) file.

WARMAC.LST added listing file to repository