robots, tell command, and locking
at the moment things work well enough if the robots don't use the 'tell' command. this suggests, focus on what's happening inside the 'tell' command and try to make things work well enough with it active as well. things have been simplified to the point where they can work ok, but it's on the edge. in fact, the 'tell' command can be active as long as there's not 'many' robots. it's both the 'tell' commmand and many robots that push it over the edge. this suggests that 'system load' is involved, whether via number of robots, frequency at which they send commands, types of commands, etc.

meanwhile, the locking code is fundamental, so also simplify there as much as possible and try to make it as stripped down and well understood. if locking can possibly work without a particular line of code, comment it out. if not forced to bring it back, leave it out.