gen2 robot
adds sophistication to be more practical for development, debugging, testing, etc. once started, pressing the '0' key will exit gracefully from decwar, tops10, and telnet, leaving you at the bash cli. it's using 'sshkeyboard' library to listen to the keyboard, and trying to be robust about quitting. meanwhile, in game, it's trying to be smarter in its command loop. for commands where it makes sense, the in game response should be auto parsed to python primitives.